functions/ingest/requirements.txt
functions-framework==3.8.2
google-cloud-bigquery==3.27.0
google-cloud-secret-manager==2.22.0
google-cloud-firestore==2.19.0
requests==2.32.3
google-cloud-logging==3.14.0